requirements.txt
flask==2.0.1
rembg==1.0.28
Pillow==8.3.1
flask-cors==3.1.1
